refactor(tab-item): move tab-item in tabs component folder
